tables and apps successfully merged
