+++ Habit_Tracker/People/apps.py
@@ -4,3 +4,7 @@ from django.apps import AppConfig
 class PeopleConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'People'
+
+class RewardsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'Rewards'
