--- Habit_Tracker/Rewards/serializers.py
@@ -1,8 +1,0 @@
-from Rewards.models import Rewards
-from rest_framework import serializers
-
-class RewardsSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = Rewards
-        fields = ['rewards_name', 'rewards_description', 'rewards_cost']
-
